update tests and use mock module
