+++ pallet-chainlink-feed/src/tests.rs
@@ -1,211 +1,6 @@
 use super::*;
-
-use frame_support::weights::Weight;
-use frame_support::{assert_noop, assert_ok, impl_outer_origin, parameter_types};
-use sp_core::H256;
-
-use frame_system as system;
-use sp_runtime::{
-	testing::Header,
-	traits::{BlakeTwo256, IdentityLookup},
-	Perbill,
-};
-
-impl_outer_origin! {
-	pub enum Origin for Test {}
-}
-
-// Configure a mock runtime to test the pallet.
-
-#[derive(Clone, Eq, PartialEq)]
-pub struct Test;
-parameter_types! {
-	pub const BlockHashCount: u64 = 250;
-	pub const MaximumBlockWeight: Weight = 1024;
-	pub const MaximumBlockLength: u32 = 2 * 1024;
-	pub const AvailableBlockRatio: Perbill = Perbill::from_percent(75);
-}
-
-type AccountId = u64;
-type BlockNumber = u64;
-impl system::Trait for Test {
-	type BaseCallFilter = ();
-	type Origin = Origin;
-	type Call = ();
-	type Index = u64;
-	type BlockNumber = BlockNumber;
-	type Hash = H256;
-	type Hashing = BlakeTwo256;
-	type AccountId = AccountId;
-	type Lookup = IdentityLookup<Self::AccountId>;
-	type Header = Header;
-	type Event = ();
-	type BlockHashCount = BlockHashCount;
-	type MaximumBlockWeight = MaximumBlockWeight;
-	type DbWeight = ();
-	type BlockExecutionWeight = ();
-	type ExtrinsicBaseWeight = ();
-	type MaximumExtrinsicWeight = MaximumBlockWeight;
-	type MaximumBlockLength = MaximumBlockLength;
-	type AvailableBlockRatio = AvailableBlockRatio;
-	type Version = ();
-	type PalletInfo = ();
-	type AccountData = pallet_balances::AccountData<u64>;
-	type OnNewAccount = ();
-	type OnKilledAccount = ();
-	type SystemWeightInfo = ();
-}
-type System = frame_system::Module<Test>;
-
-parameter_types! {
-	pub const ExistentialDeposit: u64 = 1;
-}
-
-type Balance = u64;
-impl pallet_balances::Trait for Test {
-	type MaxLocks = ();
-	type Balance = Balance;
-	type Event = ();
-	type DustRemoval = ();
-	type ExistentialDeposit = ExistentialDeposit;
-	type AccountStore = System;
-	type WeightInfo = ();
-}
-type Balances = pallet_balances::Module<Test>;
-
-const MIN_RESERVE: u64 = 100;
-
-parameter_types! {
-	pub const FeedModuleId: ModuleId = ModuleId(*b"linkfeed");
-	pub const MinimumReserve: u64 = MIN_RESERVE;
-	pub const StringLimit: u32 = 15;
-	pub const OracleLimit: u32 = 10;
-	pub const FeedLimit: u16 = 10;
-	pub const PruningWindow: u32 = 3;
-}
-
-type FeedId = u16;
-type Value = u64;
-
-impl Trait for Test {
-	type Event = ();
-	type FeedId = FeedId;
-	type Value = Value;
-	type Currency = Balances;
-	type ModuleId = FeedModuleId;
-	type MinimumReserve = MinimumReserve;
-	type StringLimit = StringLimit;
-	type OracleCountLimit = OracleLimit;
-	type FeedLimit = FeedLimit;
-	type PruningWindow = PruningWindow;
-	type WeightInfo = ();
-}
-type ChainlinkFeed = crate::Module<Test>;
-
-#[derive(Debug, Default)]
-struct FeedBuilder {
-	owner: Option<AccountId>,
-	payment: Option<Balance>,
-	timeout: Option<BlockNumber>,
-	value_bounds: Option<(Value, Value)>,
-	min_submissions: Option<u32>,
-	description: Option<Vec<u8>>,
-	restart_delay: Option<RoundId>,
-	oracles: Option<Vec<(AccountId, AccountId)>>,
-}
-
-impl FeedBuilder {
-	fn new() -> Self {
-		Self::default()
-	}
-
-	fn owner(mut self, o: AccountId) -> Self {
-		self.owner = Some(o);
-		self
-	}
-
-	fn payment(mut self, p: Balance) -> Self {
-		self.payment = Some(p);
-		self
-	}
-
-	fn timeout(mut self, t: BlockNumber) -> Self {
-		self.timeout = Some(t);
-		self
-	}
-
-	fn value_bounds(mut self, min: Value, max: Value) -> Self {
-		self.value_bounds = Some((min, max));
-		self
-	}
-
-	fn min_submissions(mut self, m: u32) -> Self {
-		self.min_submissions = Some(m);
-		self
-	}
-
-	fn description(mut self, d: Vec<u8>) -> Self {
-		self.description = Some(d);
-		self
-	}
-
-	fn restart_delay(mut self, d: RoundId) -> Self {
-		self.restart_delay = Some(d);
-		self
-	}
-
-	fn oracles(mut self, o: Vec<(AccountId, AccountId)>) -> Self {
-		self.oracles = Some(o);
-		self
-	}
-
-	fn build_and_store(self) -> DispatchResultWithPostInfo {
-		let owner = Origin::signed(self.owner.unwrap_or(1));
-		let payment = self.payment.unwrap_or(20);
-		let timeout = self.timeout.unwrap_or(1);
-		let value_bounds = self.value_bounds.unwrap_or((1, 1_000));
-		let min_submissions = self.min_submissions.unwrap_or(2);
-		let decimals = 5;
-		let description = self.description.unwrap_or(b"desc".to_vec());
-		let oracles = self.oracles.unwrap_or(vec![(2, 4), (3, 4), (4, 4)]);
-		let restart_delay = self
-			.restart_delay
-			.unwrap_or(oracles.len().saturating_sub(1) as u32);
-		ChainlinkFeed::create_feed(
-			owner,
-			payment,
-			timeout,
-			value_bounds,
-			min_submissions,
-			decimals,
-			description,
-			restart_delay,
-			oracles,
-		)
-	}
-}
-
-pub(crate) fn new_test_ext() -> sp_io::TestExternalities {
-	let mut t = frame_system::GenesisConfig::default()
-		.build_storage::<Test>()
-		.unwrap();
-
-	let module_account: AccountId = FeedModuleId::get().into_account();
-	pallet_balances::GenesisConfig::<Test> {
-		balances: vec![(module_account, 100 * MIN_RESERVE)],
-	}
-	.assimilate_storage(&mut t)
-	.unwrap();
-
-	crate::GenesisConfig::<Test> {
-		pallet_admin: module_account,
-		feed_creators: vec![1],
-	}
-	.assimilate_storage(&mut t)
-	.unwrap();
-
-	t.into()
-}
+use crate::{mock::*, Error};
+use frame_support::{assert_noop, assert_ok};
 
 #[test]
 fn feed_creation_should_work() {
@@ -544,7 +339,12 @@ fn change_oracles_should_work() {
 			),
 			Error::<Test>::OwnerCannotChangeAdmin
 		);
-		let many_duplicates: Vec<AccountId> = initial_oracles.iter().cloned().chain(initial_oracles.iter().cloned()).map(|(o, _a)| o).collect();
+		let many_duplicates: Vec<AccountId> = initial_oracles
+			.iter()
+			.cloned()
+			.chain(initial_oracles.iter().cloned())
+			.map(|(o, _a)| o)
+			.collect();
 		assert_noop!(
 			ChainlinkFeed::change_oracles(
 				Origin::signed(owner),
@@ -566,7 +366,9 @@ fn change_oracles_should_work() {
 		);
 
 		{
-			assert_ok!(ChainlinkFeed::feed_mut(feed_id).unwrap().request_new_round(AccountId::default()));
+			assert_ok!(ChainlinkFeed::feed_mut(feed_id)
+				.unwrap()
+				.request_new_round(AccountId::default()));
 		}
 		// successfully change oracles
 		assert_ok!(ChainlinkFeed::change_oracles(
@@ -749,16 +551,14 @@ fn admin_transfer_should_work() {
 			.build_and_store());
 
 		let new_admin = 42;
-		assert_noop!(ChainlinkFeed::transfer_admin(
-			Origin::signed(old_admin),
-			123,
-			new_admin
-		), Error::<Test>::OracleNotFound);
-		assert_noop!(ChainlinkFeed::transfer_admin(
-			Origin::signed(123),
-			oracle,
-			new_admin
-		), Error::<Test>::NotAdmin);
+		assert_noop!(
+			ChainlinkFeed::transfer_admin(Origin::signed(old_admin), 123, new_admin),
+			Error::<Test>::OracleNotFound
+		);
+		assert_noop!(
+			ChainlinkFeed::transfer_admin(Origin::signed(123), oracle, new_admin),
+			Error::<Test>::NotAdmin
+		);
 		assert_ok!(ChainlinkFeed::transfer_admin(
 			Origin::signed(old_admin),
 			oracle,
@@ -766,14 +566,14 @@ fn admin_transfer_should_work() {
 		));
 		let oracle_meta = ChainlinkFeed::oracle(oracle).expect("oracle should be present");
 		assert_eq!(oracle_meta.pending_admin, Some(new_admin));
-		assert_noop!(ChainlinkFeed::accept_admin(
-			Origin::signed(new_admin),
-			123,
-		), Error::<Test>::OracleNotFound);
-		assert_noop!(ChainlinkFeed::accept_admin(
-			Origin::signed(123),
-			oracle
-		), Error::<Test>::NotPendingAdmin);
+		assert_noop!(
+			ChainlinkFeed::accept_admin(Origin::signed(new_admin), 123,),
+			Error::<Test>::OracleNotFound
+		);
+		assert_noop!(
+			ChainlinkFeed::accept_admin(Origin::signed(123), oracle),
+			Error::<Test>::NotPendingAdmin
+		);
 		assert_ok!(ChainlinkFeed::accept_admin(
 			Origin::signed(new_admin),
 			oracle
@@ -827,15 +627,15 @@ fn request_new_round_should_work() {
 			}
 		);
 		// failure cases
-		assert_noop!(ChainlinkFeed::request_new_round(
-			Origin::signed(123),
-			feed_id
-		), Error::<Test>::NotAuthorizedRequester);
+		assert_noop!(
+			ChainlinkFeed::request_new_round(Origin::signed(123), feed_id),
+			Error::<Test>::NotAuthorizedRequester
+		);
 		// non existent feed is not present but also not authorized
-		assert_noop!(ChainlinkFeed::request_new_round(
-			Origin::signed(requester),
-			123
-		), Error::<Test>::NotAuthorizedRequester);
+		assert_noop!(
+			ChainlinkFeed::request_new_round(Origin::signed(requester), 123),
+			Error::<Test>::NotAuthorizedRequester
+		);
 
 		// actually request new round
 		assert_ok!(ChainlinkFeed::request_new_round(
@@ -843,16 +643,16 @@ fn request_new_round_should_work() {
 			feed_id
 		));
 		// need to wait `delay` rounds before requesting again
-		assert_noop!(ChainlinkFeed::request_new_round(
-			Origin::signed(requester),
-			feed_id
-		), Error::<Test>::CannotRequestRoundYet);
+		assert_noop!(
+			ChainlinkFeed::request_new_round(Origin::signed(requester), feed_id),
+			Error::<Test>::CannotRequestRoundYet
+		);
 		// round does not have data and is not timed out
 		// --> not supersedable
-		assert_noop!(ChainlinkFeed::request_new_round(
-			Origin::signed(snd_requester),
-			feed_id
-		), Error::<Test>::RoundNotSupersedable);
+		assert_noop!(
+			ChainlinkFeed::request_new_round(Origin::signed(snd_requester), feed_id),
+			Error::<Test>::RoundNotSupersedable
+		);
 
 		let round_id = 1;
 		let round = ChainlinkFeed::round(feed_id, round_id).expect("first round should be present");
@@ -1079,7 +879,8 @@ fn payment_withdrawal_should_work() {
 		let fund_balance = Balances::free_balance(&fund);
 		Balances::make_free_balance_be(&fund, ExistentialDeposit::get());
 		assert!(
-			ChainlinkFeed::withdraw_payment(Origin::signed(admin), oracle, recipient, amount).is_err()
+			ChainlinkFeed::withdraw_payment(Origin::signed(admin), oracle, recipient, amount)
+				.is_err()
 		);
 		Balances::make_free_balance_be(&fund, fund_balance);
 
@@ -1123,18 +924,19 @@ fn transfer_pallet_admin_should_work() {
 	new_test_ext().execute_with(|| {
 		let new_admin = 23;
 		let fund = FeedModuleId::get().into_account();
-		assert_noop!(ChainlinkFeed::transfer_pallet_admin(
-			Origin::signed(123),
-			new_admin
-		), Error::<Test>::NotPalletAdmin);
+		assert_noop!(
+			ChainlinkFeed::transfer_pallet_admin(Origin::signed(123), new_admin),
+			Error::<Test>::NotPalletAdmin
+		);
 		assert_ok!(ChainlinkFeed::transfer_pallet_admin(
 			Origin::signed(fund),
 			new_admin
 		));
 		assert_eq!(PendingPalletAdmin::<Test>::get(), Some(new_admin));
-		assert_noop!(ChainlinkFeed::accept_pallet_admin(
-			Origin::signed(123)
-		), Error::<Test>::NotPendingPalletAdmin);
+		assert_noop!(
+			ChainlinkFeed::accept_pallet_admin(Origin::signed(123)),
+			Error::<Test>::NotPendingPalletAdmin
+		);
 		assert_ok!(ChainlinkFeed::accept_pallet_admin(Origin::signed(
 			new_admin
 		)));
@@ -1266,10 +1068,7 @@ fn feed_creation_permissioning() {
 			Error::<Test>::NotFeedCreator
 		);
 		assert_noop!(
-			ChainlinkFeed::set_feed_creator(
-				Origin::signed(123),
-				new_creator
-			),
+			ChainlinkFeed::set_feed_creator(Origin::signed(123), new_creator),
 			Error::<Test>::NotPalletAdmin
 		);
 		assert_ok!(ChainlinkFeed::set_feed_creator(
@@ -1278,10 +1077,7 @@ fn feed_creation_permissioning() {
 		));
 		assert_ok!(FeedBuilder::new().owner(new_creator).build_and_store());
 		assert_noop!(
-			ChainlinkFeed::remove_feed_creator(
-				Origin::signed(123),
-				new_creator
-			),
+			ChainlinkFeed::remove_feed_creator(Origin::signed(123), new_creator),
 			Error::<Test>::NotPalletAdmin
 		);
 		assert_ok!(ChainlinkFeed::remove_feed_creator(
@@ -1357,7 +1153,12 @@ fn feed_life_cylce() {
 		{
 			let mut feed = Feed::<Test>::new(id, new_config.clone());
 			assert_ok!(feed.add_oracles(oracles.clone()));
-			assert_ok!(feed.update_future_rounds(payment, submission_count_bounds, restart_delay, timeout));
+			assert_ok!(feed.update_future_rounds(
+				payment,
+				submission_count_bounds,
+				restart_delay,
+				timeout
+			));
 		}
 		let new_config = FeedConfig {
 			oracle_count: oracles.len() as u32,
@@ -1375,14 +1176,20 @@ fn feed_life_cylce() {
 			..new_config.clone()
 		};
 		// modified config should be stored on drop
-		assert_eq!(ChainlinkFeed::feed_config(id), Some(modified_config.clone()));
+		assert_eq!(
+			ChainlinkFeed::feed_config(id),
+			Some(modified_config.clone())
+		);
 		let ignored_timeout = 23;
 		{
 			let mut feed = Feed::<Test>::read_only_from(id).expect("feed should be there");
 			feed.config.timeout = ignored_timeout;
 		}
 		// read only access should not store changes
-		assert_eq!(ChainlinkFeed::feed_config(id), Some(modified_config.clone()));
+		assert_eq!(
+			ChainlinkFeed::feed_config(id),
+			Some(modified_config.clone())
+		);
 		{
 			let mut feed = ChainlinkFeed::feed_mut(id).expect("feed should be there");
 			assert_ok!(feed.request_new_round(AccountId::default()));
